Renamed package to reflect that this is more than just samples.
