+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/sandbox/PedroAutoSample.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.samples;
+package org.firstinspires.ftc.teamcode.sandbox;
 
 import com.pedropathing.follower.Follower;
 import com.pedropathing.geometry.BezierCurve;
