+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/sandbox/PedroTeleOpSample.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.samples;
+package org.firstinspires.ftc.teamcode.sandbox;
 
 import com.pedropathing.follower.Follower;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
